For go modules to work we needed to change the directory of the application to github.com/stablyprime. The go.mod was re-initialized for the the new module name.
Added a handler to process common base code for an incoming request. The idea is to allow the project to be run as a function or as a complete
service either on premise or in a virtual instance in the cloud (i.e EC2). As the parameters vary from one request to another we are
establishing this handler to allow the application to process different requests to grab required request parameters and use the same logic
without caring who is making the request.

All the logic that is not related to the request itself is being placed in an individual api, this will allow us to test the api without caring
about input http request parameters. This would also apply if we would want to read from the database, we could interface a db connection or a mock
connection.